\name{prot}
\alias{prot}
\docType{data}
\title{Proteomic Plasma Preparation Record.}
\description{
Proteomic record. See Donnie Stapleton.
}
\usage{data(prot)}
\format{
  A data frame with 604 observations on the following 14 variables.
  \describe{
    \item{\code{Strain}}{a factor with levels \code{B6} \code{BTBR} \code{F1} \code{F2}}
    \item{\code{MouseNum}}{a numeric vector}
    \item{\code{Tube}}{Tube ID.}
    \item{\code{X}}{a factor with levels \code{Box 1} \code{Box 2} \code{Box 3} \code{Box 4} \code{Box 5} \code{Box 6} \code{Box 7} \code{Box 8}}
    \item{\code{X.1}}{a numeric vector}
    \item{\code{ul.of.plasma.remaining}}{a numeric vector}
    \item{\code{CPL.UC.Davis}}{a numeric vector}
    \item{\code{RBM}}{a numeric vector}
    \item{\code{X.50}}{a numeric vector}
    \item{\code{X.100}}{a numeric vector}
    \item{\code{X.150}}{a numeric vector}
    \item{\code{X.200}}{a numeric vector}
    \item{\code{X.300}}{a numeric vector}
    \item{\code{X.400}}{a numeric vector}
  }
}
\examples{
data(prot)
}
\keyword{datasets}
